video/huc6272.cpp: convert KRAM data access to 16-bit

+++ src/devices/video/huc6272.h
@@ -95,8 +95,8 @@ private:
 	const address_space_config      m_data_space_config;
 	const address_space_config      m_io_space_config;
 	required_shared_ptr<uint16_t>   m_microprg_ram;
-	required_shared_ptr<uint32_t>   m_kram_page0;
-	required_shared_ptr<uint32_t>   m_kram_page1;
+	required_shared_ptr<uint16_t>   m_kram_page0;
+	required_shared_ptr<uint16_t>   m_kram_page1;
 	required_device<scsi_port_device> m_scsibus;
 	required_device<input_buffer_device> m_scsi_data_in;
 	required_device<output_latch_device> m_scsi_data_out;
@@ -142,8 +142,8 @@ private:
 	u32 kram_write_address_r(offs_t offset);
 	void kram_write_address_w(offs_t offset, u32 data, u32 mem_mask = ~0);
 
-	u32 kram_read_data_r(offs_t offset, u32 mem_mask = ~0);
-	void kram_write_data_w(offs_t offset, u32 data, u32 mem_mask = ~0);
+	u16 kram_read_data_r(offs_t offset, u16 mem_mask = ~0);
+	void kram_write_data_w(offs_t offset, u16 data, u16 mem_mask = ~0);
 
 	u32 kram_page_setup_r(offs_t offset);
 	void kram_page_setup_w(offs_t offset, u32 data, u32 mem_mask = ~0);
